Introduce explicit 'PartialPayment' transaction:
The XRP Ledger supports powerful payment functionality and several
advanced features. One of those features is the ability to send
payments which are permitted to deliver less than the amount
originally specified in the transaction.

Such payments are both common and useful. By way of example, consider
credit and debit card payments: the customer is quoted a price and
her account is debited exactly the quoted amount. However, the amount
that will actually be credited to the merchant's account will be less
because of, inter alia, transaction fees imposed by the credit card
processor.

Ripple presently supports such payments by tagging payment transactions
with the `tfPartialPayment` flag. The amount that is actually delivered
by such payments may be less than the amount that is specified in the
`Amount` field, and is reported in the transaction's metadata as
`DeliveredAmount`.

This commit introduces the new "PartialPayments" amendment, making
the partial nature of the payment explicit as opposed to conditioned
on the presence of the flag.

If this amendment is activated, it will:

- Remove support for the `tfPartialPayment` flag. Any new payment
  transaction submitted with this flag set will fail with a
  `temINVALID_FLAG` error.
- Introduce a new `PartialPayment` transaction type which can be
  used to execute a partial payment. The `Payment` transaction will
  continue to be used to process regular payments.

As always, to determine the amount that a payment actually delivers
the `DeliveredAmount` field in the transaction's metadata should be
consulted.