Allow account holders to disable incoming partial payments:
The XRP Ledger supports several advanced features. One of those features
is the ability to send "partial payments": payments which are permitted
to deliver less than the amount originally specified in the transaction.

Such payments are both common and useful. By way of example, consider
credit and debit card payments: the customer is quoted a price and
her account is debited exactly the quoted amount. However, the amount
that will actually be credited to the merchant's account will be less
because of, inter alia, transaction fees imposed by the credit card
processor.

This commit introduces a new amendment `BlockPartialPayments` which, if
activated will:

- Introduce a new account flag which, if set, will prevent the account
  from receiving payments with the `tfPartialPayment` flag set even
  if such payments would have delivered the full amount requests.
  This flag does not affect partial payments to oneself, which will
  continue to be allowed.

- Make the new account flag the default for all new wallets.

As always, to determine the amount that a payment actually delivered
you should disregard the `Amount` field in the transaction and always
consult the `DeliveredAmount` field which is present in the transaction's
metadata.