[WIP] Improve hex codec and reduce custom code:
We currently have several instances of hand-rolled hex encoding and
decoding routines. It would be great if we could reduce the amount
of custom code by leveraging an existing library and reduce the
number of entry poins by leveraging template metaprogramming.

This commit attempts to do this by introducing two new interfaces:
to_hex and from_hex which are based on Boost.Algorithms and replace
the various existing routines.

There is a small change in behavior: our existing code would allow
converting an odd-length hex string by assuming the leading byte
was a 0.