Introduce support for deletable accounts:
The XRP Ledger utilizes an account model. Unlike systems based on a UTXO
model, XRP Ledger accounts are first-class objects. This design choice
allows the XRP Ledger to offer rich functionality, including the ability
to own objects (offers, escrows, checks, signer lists) as well as other
advanced features, such as key rotation and configurable multi-signing
without needing to change a destination address.

The trade-off is that accounts must be stored on ledger. The XRP Ledger
applies reserve requirements, in XRP, to protect the shared global ledger
from growing excessively large as the result of spam or malicious usage.

Prior to this tweet, accounts had been permanent objects; once created,
they could never be deleted.

This commits introduces a new amendment "DeletableAccounts" which, if
enabled, will allow account objects to be deleted by executing the new
"DeleteAccount" transaction. Any funds remaining in the account will
be transferred to an specified in the deletion transaction.

The amendment changes the mechanics of account creation; previously
new account would have an initial sequence number of 1. Accounts
created after the amendment will have an initial sequence number that
is equal to the ledger in which the account was created.

Accounts can only be deleted if they do not own any objects and if the
current ledger sequence number exceeds the account's sequence number so
that, if recreated, the account can be protected from transaction replay.

The amendment also introduces a new flag, that will be set by default
on all new accounts. The flag indicates that the account does not plan
to issue on-ledger IOUs and prevents others from opening trust lines to
that account. This flag is used to prevent a race condition where an
account could be made "not deletable" by having a trust line opened.