[WIP] Harden validations:
This commit introduces a new optional field in validation messages
which, if set, will allow a validator to attest to the hash of the
latest ledger that it considers to be fully validated.

This commit also leverages the previously introduced "cookie" field
to improve the robustness of the network by making it possible for
servers to automatically detect accidental misconfiguration which
results in two or more validators using the same validation key.