Rework deferred node logic and async fetch behavior:
This change significantly improves ledger sync and fetch times
while reducing memory consumption. The change affects the code
from that begins with SHAMap::getMissingNodes and runs through
to Database::threadEntry.

Prior to this commit, the code would issue asynchronous fetches
which were handed off to the Database's pool of read threads for
execution and the results of each read would be placed in the
Database's positive and negative caches. The caller would have
to wait for all reads to complete and then retrieve the results
out of these caches.

Among other issues, this means that the results of the first read
cannot be processed until the last read completes. Additionally,
all the results must sit in memory.

This commit changes the behavior so that each read operation has a
completion handler associated with it. The completion of the read
calls the handler, so the results of each read can be processed
as it completes, eliminating the need for the both the negative
and positive caches.

Additionally, this commit also removes the read generation code
which is also no longer needed, and the vestigial batch fetch
logic which was never fully implemented nor supported.